test: separate test cpp file and change to parser
